requirements.txt
streamlit
seaborn
pandas
openai
python-dotenv
google-generativeai